mitosheet: fix regex use in split text to columns and filter
