+++ mitosheet/mitosheet/saved_analyses/schema_utils.py
@@ -73,23 +73,3 @@ def upgrade_saved_analysis_format_to_steps_data(saved_analysis: Optional[Dict[st
         }
     else:
         return saved_analysis
-
-
-
-def is_prev_version(curr_version: str, benchmark_version: str) -> bool:
-    """
-    Returns True if the curr_version is previous to the benchmark_version
-    Note that this assumes semantic versioning with x.y.z!
-    """
-    curr_version_parts = curr_version.split('.')
-    benchmark_version_parts = benchmark_version.split('.')
-
-    for old_version_part, benchmark_version_part in zip(curr_version_parts, benchmark_version_parts):
-        if int(old_version_part) > int(benchmark_version_part):
-            # E.g. if we have 0.2.11 and 0.1.11, we want to return early as it's clearly not older!
-            return False
-
-        if int(old_version_part) < int(benchmark_version_part):
-            return True
-
-    return False
