+++ mitosheet/mitosheet/step_performers/column_steps/split_text_to_columns.py
@@ -52,11 +52,11 @@ class SplitTextToColumnsStepPerformer(StepPerformer):
 
         # Create the dataframe of new columns. We do this first, so that we know how many columns get created.
         if is_datetime_dtype(str(post_state.dfs[sheet_index][column_header].dtype)):
-            new_columns_df = final_df[column_header].dt.strftime('%Y-%m-%d %X').str.split(delimiter_string, -1, **split_param_dict)
+            new_columns_df = final_df[column_header].dt.strftime('%Y-%m-%d %X').str.split(delimiter_string, **split_param_dict)
         elif is_timedelta_dtype(str(post_state.dfs[sheet_index][column_header].dtype)):
-            new_columns_df = final_df[column_header].apply(lambda x: str(x)).str.split(delimiter_string, -1, **split_param_dict)
+            new_columns_df = final_df[column_header].apply(lambda x: str(x)).str.split(delimiter_string, **split_param_dict)
         else:
-            new_columns_df = final_df[column_header].astype('str').str.split(delimiter_string, -1, **split_param_dict)
+            new_columns_df = final_df[column_header].astype('str').str.split(delimiter_string, **split_param_dict)
 
         # Create the new column headers and ensure they are unique
         # Note: We create the new_column_header_suffix on the frontend so that it is saved in the step parameters, which allows us
