+++ mitosheet/mitosheet/transpiler/transpile_utils.py
@@ -9,6 +9,7 @@ from typing import Any, Dict, List, Optional, Set, Tuple, Union
 import pandas as pd
 import numpy as np
 from mitosheet.types import ColumnHeader
+from mitosheet.utils import is_prev_version
 
 # TAB is used in place of \t in generated code because
 # Jupyter turns \t into a grey arrow, but converts four spaces into a tab.
@@ -45,7 +46,6 @@ def column_header_to_transpiled_code(column_header: ColumnHeader) -> str:
         return f'({column_header_parts_joined})'
 
     # We must handle np.nan first because isinstance(np.nan, float) evaluates to True
-    from mitosheet.saved_analyses.schema_utils import is_prev_version
     if not is_prev_version(pd.__version__, '1.0.0') and column_header is np.nan:
         return 'pd.np.nan'
     elif isinstance(column_header, int) or isinstance(column_header, float) or isinstance(column_header, bool):
